Changed OnColorSchemeChanged to be private
This is a static method that's used internally and has no use or purpose to be public. I marked these as public previously simply because younger me didn't understand that these could've instead been private.

While this is a breaking change technically, this will be a change that prevents bad or brainless behavior from occurring.