Fix parameter names, add note about Select method
Fix the parameter names for a few things implementing IClickSelectableControl, and SelectableCollectionView.

Also wrote a note about how the Select method in ISelectableCollectionSource will be renamed to SelectItem, to avoid collision with Linq's Select and to fix the code warning from the analyzer.

This should be doable to port forward to 2.0.