Use provided culture in XAML value converters
The nice thing with the value converters is that they already have the culture info provided as a parameter, that I can pass along to the methods inside. Makes these easy to fix the analyzer issues for.

Some of these have slightly other issues though that I also solved. I also improved XML comments for DoubleStringConverter.

This could be a candidate to port forward to 2.0, but since I moved all the converter files to another folder, I'm not sure if Git will pick that up? Maybe I should manually copy this code over.