Explicity set culture to null to use current culture
This fixes a few more of these code warnings in these places. Technically, behavior hasn't changed, but now I'm being more explicit in my code that the current culture is being used (for which, "null" is a shorthand for). (Although for ColorListItem, I use InvariantCulture, since I want that experience to be the same in all situations.)

This could be ported forward to 2.0.