Fix analyzer warnings for property list editors
For these, explicity set these to use the current culture info in most places. Also set the string comparison to ordinal for doing things like string.StartsWith.

I'm wondering for the future if I should provide a property or method to indicate what culture or info to use for the property list editors, like WPF does for the value converters. I'll think about it.