Fix menu item brushes for Menu, ContextMenu
Rather than menu items by default using the brushes that are applied to the background, instead they use a almost-fully-transparent brush that still allows them to be hit tested.

This prevents issues where a gradient brush is used for the menu's background and it ends up applied to the menu items too, ending up with weird results.