Spinners: dispose timers when unloading
This solves the analyzer warning about using disposable fields. I also added in protected methods to these, ResetTimers, that can be used in case a more advanced scenario is needed (such as the control being unloaded and then later loaded again). In that case, a developer could inherit a spinner control and use that method.

This might be forward-port-able to 2.0? Maybe?

Anyway, final thing done, time for release... tomorrow!