Adjust SsuiTheme's HighlightForeground color
An attempt to improve the visuals of the CheckBox to be closer to what I had intended for 2.0. This updates the CheckHighlightBrush to use the HighlightForeground, and then changes what HighlightForeground will look like. In general, I think this HighlightForeground change is good beyond even just the CheckBox, but it didn't fully work out the way I wanted it to for CheckBox.

I think what I will need to do is create a CheckHighlight property, which I'm not 100% keen on, to be honest, but I think will be needed.