Add attributes to properties, various property fixes
Went through a handful of controls to add in attributes to make sure that the category is set, and also add on a Description category too. Apparently Description is what the WPF designer uses for displaying info in the tooltip, unlike the rest of Visual Studio seemingly using the XML comments. I also added the ReadOnly attribute to a selection of get-only properties (only the ones that I'm not expecting the values' children to be edited). Not sure if that does anything, but why not, I guess?

During this, also spent a decent bit of time cleaning up and improving various bits of code related to dependency properties too. Biggest things are adding more usage of nameof, and rearranging the categories that things are in.

In a future commit, I'll have to go through more of the controls (in particular, CheckBox, most of the spinner controls, PositionSelect, MenuButton, LinkTextBlock) to add in the attributes to these properties too.