Improvements to MessageDialog
Renamed some of the parameters for ShowDialog, and some of the properties (in particular to rename the vaguely-named "extra" buttons to "choice" buttons; still not the most specific name, but at least it makes a tiny bit more sense).

Added new default action/result functionality, since the IsDefault functionality didn't work for buttons. Now the default button will be highlighted/focused when the dialog is opened.

In the future, I may do more, especially with handling key presses. I added a handler for the Escape key, but I'd like to add more.