Change CaptionButtonPadding default to 15,7
Just going to do it now for SSUI 2.0, to get it adjusted to the Windows 10/11 style, rather than doing the weird middle step that I was previously planning.

It's a major version number increase, so let's just break as much as I can, I guess!

This is a breaking change, I guess lol.