Update appearance of TabDisplayItem's close button
Was originally looking at doing a look that removed the border altogether from the Close button, but it turns out that by default with the SsuiTheme, the tab's highlight brush is the same as the button's highlight brush, and so the two aren't distinguishable. Not sure if I want to change that, but if I do, what would I change it to?

Instead, I just made some minor adjustments that tried to remove having the extra borders where it wasn't needed. It is a bit debatable how much of an improvement that really is, but meh. I also had the button work better at smaller tab bar heights.

Finally, I also simplified the MenuItem template style used in TabControl for the Tab List Menu. Makes that code not as large to scroll through lol.