Add ThemeValueExclude property to ThemedControl
This is the solution I came up with for situations where you dont' want your brushes/values set via XAML to be overwritten when applying the SsuiTheme.

From initial testing, it seems to work! But yet, not quite the way I was expecting/wanting with SplitButton... so I'll have to see what's going on with that. I'm worried that I might have to specially handle CornerRadius via C# for SplitButton, which would not be fun.

Anyway, this is a major part done, and I think that mostly wraps up everything I want to do with ThemedControl and similar classes!