Attempts to fix binding/brushes for SplitButton
The attempts didn't work lol. So now it's even more broken right now. But I think I have a better basis to build from. I'm going to have to do the annoying code-behind applying of brushes/properties that I had to do with the buttons in the TabControl, aren't I?